Refactor: 스타일 코드 분리 및 내보내기 방식 변경 (#226)
* #224 refactor: 스타일 코드 분리 및 Barrel Export 패턴 적용

- 스타일 코드가 분리된 컴포넌트에 한해 폴더 단위로 정리

* #224 refactor: Sidebar를 Default Export로 변경

- 불필요한 aria-label 속성 제거
