+++ src/components/Notification/index.tsx
@@ -2,17 +2,24 @@ import { hasUnreadNotifications } from '@/apis';
 import { Check, Close } from '@/assets/icons';
 import { IconWrapper } from '@/components';
 import { Button, Overlay } from '@/components/common';
-import { commonButtonStyle, headerStyle, sidebarContainer } from '@/components/common/Sidebar';
+import { commonButtonStyle, headerStyle, sidebarContainer } from '@/components/common/Sidebar/style';
+import {
+  acceptIconStyle,
+  buttonStyle,
+  checkIconStyle,
+  itemStyle,
+  listStyle,
+  markStyle,
+} from '@/components/Notification/style';
 import { INVITATION_TYPE } from '@/constants/squad';
 import { useNotificationUpdateContext } from '@/context/notification';
 import { useInfinite, useToastHandler } from '@/hooks';
 import { useAcceptInvitation } from '@/hooks/mutations';
 import useUpdateNotification from '@/hooks/mutations/useUpdateNotification';
 import { notificationInfiniteQueryOptions, squadKeys } from '@/hooks/queries';
-import { fullSizeButtonStyle, scrollBarStyle } from '@/styles/globalStyles';
+import { fullSizeButtonStyle } from '@/styles/globalStyles';
 import { NotificationResponse, NotificationUpdateRequestPayload } from '@/types';
 import { getDateRange } from '@/utils/getDateRange';
-import { css, Theme } from '@emotion/react';
 import { useInfiniteQuery, useQueryClient } from '@tanstack/react-query';
 import { MouseEvent, MouseEventHandler } from 'react';
 
@@ -151,66 +158,3 @@ const AcceptStatus = ({ isAccept, onAccept }: { isAccept: boolean; onAccept: Mou
 
 Notification.List = NotificationList;
 Notification.Item = NotificationItem;
-
-const listStyle = css`
-  margin-top: 24px;
-  overflow-y: auto;
-
-  ${scrollBarStyle}
-`;
-
-const itemStyle = (theme: Theme) => css`
-  height: 56px;
-  display: flex;
-  align-items: flex-start;
-  padding: 8px 12px;
-  ${theme.typography.size_14}
-  font-weight: 400;
-  flex: 2;
-
-  & p {
-    width: 100%;
-    text-align: left;
-  }
-
-  & button {
-    width: 56px;
-  }
-
-  :hover {
-    background-color: #e2e2e250;
-    cursor: pointer;
-  }
-`;
-
-const markStyle = (isRead: boolean) => css`
-  ::before {
-    opacity: ${isRead && 0};
-    content: '';
-    width: 8px;
-    height: 6px;
-    background-color: red;
-    border-radius: 50%;
-    margin-top: 8px;
-  }
-`;
-
-const buttonStyle = () => css`
-  opacity: 50%;
-  pointer-events: none;
-`;
-
-const acceptIconStyle = css`
-  width: 56px;
-  display: flex;
-  justify-content: center;
-`;
-
-const checkIconStyle = {
-  width: '24px',
-  height: '24px',
-  borderRadius: '50%',
-  backgroundColor: 'var(--color-success)',
-  color: '#fff',
-  padding: '4px',
-};
