+++ src/components/Todo/Item.tsx
@@ -4,7 +4,7 @@ import { TODO_STATUS } from '@/constants/todo';
 import { useTodoDeleteActions, useTodoUpdateActions } from '@/hooks/todo';
 import { ToDoDetail, TodoQueryParams } from '@/types';
 import { useTheme } from '@emotion/react';
-import { memo, useMemo } from 'react';
+import { memo } from 'react';
 
 const Item = ({ todo, queryParams }: { todo: ToDoDetail; queryParams: TodoQueryParams }) => {
   const theme = useTheme();
@@ -14,14 +14,12 @@ const Item = ({ todo, queryParams }: { todo: ToDoDetail; queryParams: TodoQueryP
   const { isDeleteMode, handleConfirmDelete, toggleDeleteMode } = useTodoDeleteActions(todo.toDoId, queryParams);
   const todoUpdateActions = useTodoUpdateActions(todo, queryParams);
 
-  const memoizedStatusStyle = useMemo(() => [todoItemStyle, getStatusStyles(isCompleted, theme)], [isCompleted, theme]);
-
   if (isDeleteMode) {
     return <DeleteMode onChangeMode={toggleDeleteMode} onDelete={handleConfirmDelete} />;
   }
 
   return (
-    <li css={memoizedStatusStyle} aria-labelledby={`todo-${todo.toDoId}`}>
+    <li css={[todoItemStyle, getStatusStyles(isCompleted, theme)]} aria-labelledby={`todo-${todo.toDoId}`}>
       <Content todo={todo} isCompleted={isCompleted} updateActions={todoUpdateActions} />
       <ActionButton contents={contents} updateActions={todoUpdateActions} onChangeDeleteMode={toggleDeleteMode} />
     </li>
