--- src/components/common/Button/index.ts
@@ -1,3 +1,0 @@
-import Button from '@/components/common/Button/Button';
-
-export { Button };
