--- src/components/common/Calendar/index.ts
@@ -1,3 +1,0 @@
-import Calendar from '@/components/common/Calendar/Calendar';
-
-export { Calendar };
