+++ src/components/common/ErrorBoundary/RootErrorFallback/index.tsx
@@ -1,5 +1,12 @@
+import { buttonStyle } from '@/components/common/Button/style';
+import {
+  buttonGroupStyle,
+  eggIconStyle,
+  errorMessageStyle,
+  fallbackWrapperStyle,
+  homeButton,
+} from '@/components/common/ErrorBoundary/RootErrorFallback/style';
 import { useUserCookie } from '@/hooks';
-import { css } from '@emotion/react';
 import { useQueryClient } from '@tanstack/react-query';
 import { FallbackProps } from 'react-error-boundary';
 import { useNavigate } from 'react-router-dom';
@@ -57,55 +64,3 @@ const RootErrorFallback = ({ error, resetErrorBoundary }: FallbackProps) => {
 };
 
 export default RootErrorFallback;
-
-const fallbackWrapperStyle = css`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 100vh;
-  background: #fff5e6;
-  color: #333;
-  text-align: center;
-  font-family: 'Arial', sans-serif;
-`;
-
-const errorMessageStyle = css`
-  font-size: 1.2rem;
-  line-height: 1.8;
-  max-width: 500px;
-  margin-bottom: 2rem;
-  color: #ffa502;
-  font-weight: 700;
-`;
-
-const buttonGroupStyle = css`
-  display: flex;
-  gap: 8px;
-`;
-
-const buttonStyle = css`
-  padding: 8px 16px;
-  background: #ffa502;
-  color: white;
-  font-size: 1rem;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  transition: background 0.3s ease;
-
-  &:hover {
-    background: #ffbd43;
-  }
-`;
-
-const homeButton = css`
-  background-color: white;
-  color: #ffa502;
-  border: 1px solid #ffa502;
-`;
-
-const eggIconStyle = css`
-  font-size: 5rem;
-  margin-bottom: 1rem;
-`;
