--- src/components/common/Form/index.ts
@@ -1,3 +1,0 @@
-import Form from '@/components/common/Form/Form';
-
-export { Form };
