--- src/components/common/Loading/index.ts
@@ -1,3 +1,0 @@
-import Loading from '@/components/common/Loading/Loading';
-
-export { Loading };
