+++ src/components/common/Modal/ModalContents/SquadForm/index.tsx
@@ -1,10 +1,10 @@
 import { Button } from '@/components/common';
 import { ModalButtonGroup, ModalContent, ModalHeader, ModalTemplate } from '@/components/common/Modal';
+import { descStyle, inputStyle } from '@/components/common/Modal/ModalContents/SquadForm/style';
 import { ModalContentProps } from '@/types';
-import { css, Theme } from '@emotion/react';
 import { FormEvent, useState } from 'react';
 
-export const SquadForm = ({
+const SquadForm = ({
   onSubmit,
   onAbort,
   squadName = '',
@@ -46,16 +46,4 @@ export const SquadForm = ({
   );
 };
 
-export const descStyle = (theme: Theme) => css`
-  color: ${theme.colors.black.black300};
-`;
-
-export const inputStyle = (theme: Theme) => css`
-  width: 100%;
-  height: 40px;
-  border: 1.5px solid #dfdfdf;
-  border-radius: 6px;
-  margin: 16px 0;
-  padding: 0 8px;
-  ${theme.typography.size_16}
-`;
+export default SquadForm;
