+++ src/components/common/Modal/ModalContents/index.ts
@@ -1,5 +1,5 @@
 import ActionModal from '@/components/common/Modal/ModalContents/ActionModal';
 import EmailInputModal from '@/components/common/Modal/ModalContents/EmailInputModal';
-import { SquadForm } from '@/components/common/Modal/ModalContents/SquadForm';
+import SquadForm from '@/components/common/Modal/ModalContents/SquadForm';
 
 export { ActionModal, EmailInputModal, SquadForm };
