+++ src/components/common/Modal/ModalTemplate/ModalTemplate.tsx
@@ -1,12 +1,6 @@
-import { Overlay, OverlayProps } from '@/components/common/Overlay';
-import { ModalStyle } from '@/types';
-import { PropsWithChildren, memo } from 'react';
-
-export type ModalTemplate = PropsWithChildren &
-  OverlayProps & {
-    isOverlay: boolean;
-    styleType?: ModalStyle;
-  };
+import { Overlay } from '@/components/common';
+import { ModalTemplate } from '@/types';
+import { memo } from 'react';
 
 const Container = memo(({ children, ...rest }: Omit<ModalTemplate, 'styleType' | 'title'>) =>
   rest.isOverlay ? (
