+++ src/components/common/Modal/index.ts
@@ -1,4 +1,6 @@
+import { ActionModal, EmailInputModal, SquadForm } from '@/components/common/Modal/ModalContents';
+import { ModalButtonGroup, ModalContent, ModalHeader, ModalTemplate } from '@/components/common/Modal/ModalTemplate';
+
 export { default as Modal } from './Modal';
-export * from './ModalContents';
 export { default as ModalPortal } from './ModalPortal';
-export * from './ModalTemplate';
+export { ActionModal, EmailInputModal, ModalButtonGroup, ModalContent, ModalHeader, ModalTemplate, SquadForm };
