--- src/components/common/Overlay/index.ts
@@ -1,1 +1,0 @@
-export { default as Overlay, type OverlayProps } from './Overlay';
