--- src/components/common/Sidebar/index.ts
@@ -1,3 +1,0 @@
-import { Sidebar, commonButtonStyle, headerStyle, sidebarContainer } from '@/components/common/Sidebar/Sidebar';
-
-export { Sidebar, commonButtonStyle, headerStyle, sidebarContainer };
