+++ src/components/common/Sidebar/index.tsx
@@ -1,7 +1,16 @@
 import { Close, Edit } from '@/assets/icons';
 import { IconWrapper } from '@/components';
 import { SidebarMemberList } from '@/components/Member';
-import { Overlay } from '@/components/common/Overlay';
+import { Overlay } from '@/components/common';
+import {
+  commonButtonStyle,
+  exitButtonStyle,
+  headerStyle,
+  membersStyle,
+  settingsStyle,
+  sidebarContainer,
+  squadInfoStyle,
+} from '@/components/common/Sidebar/style';
 import { ROLE } from '@/constants/role';
 import { useUserCookie } from '@/hooks';
 import { useDeleteSquad, useExitSquad, useUpdateSquadName } from '@/hooks/mutations';
@@ -9,11 +18,10 @@ import { squadDetailQueryOptions, squadKeys } from '@/hooks/queries';
 import { useSquadStore, useToastStore } from '@/stores';
 import { fullSizeButtonStyle } from '@/styles/globalStyles';
 import { getPriority } from '@/utils';
-import { css, Theme } from '@emotion/react';
 import { useQueryClient, useSuspenseQuery } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
 
-export const Sidebar = ({ closeSidebar }: { closeSidebar: VoidFunction }) => {
+const Sidebar = ({ closeSidebar }: { closeSidebar: VoidFunction }) => {
   const squadId = useSquadStore((state) => state.currentSquadId);
   const navigate = useNavigate();
   const createToast = useToastStore((state) => state.createToast);
@@ -107,20 +115,20 @@ export const Sidebar = ({ closeSidebar }: { closeSidebar: VoidFunction }) => {
           <h2 id="squad-settings">설정</h2>
           <ul>
             <li>
-              <button onClick={handleInvitation} style={fullSizeButtonStyle} aria-label="멤버 초대">
+              <button onClick={handleInvitation} style={fullSizeButtonStyle}>
                 멤버 초대
               </button>
             </li>
             {isLeader && (
               <li>
-                <button onClick={handleAssignLeader} style={fullSizeButtonStyle} aria-label="리더 변경">
+                <button onClick={handleAssignLeader} style={fullSizeButtonStyle}>
                   리더 변경
                 </button>
               </li>
             )}
             {isLeader && (
               <li>
-                <button onClick={handleSquadDelete} style={fullSizeButtonStyle} aria-label="스쿼드 삭제">
+                <button onClick={handleSquadDelete} style={fullSizeButtonStyle}>
                   스쿼드 삭제
                 </button>
               </li>
@@ -138,106 +146,4 @@ export const Sidebar = ({ closeSidebar }: { closeSidebar: VoidFunction }) => {
   );
 };
 
-export const commonButtonStyle = (theme: Theme) => css`
-  width: 28px;
-  height: 28px;
-  color: ${theme.colors.text};
-`;
-
-export const sidebarContainer = (theme: Theme) => css`
-  width: 300px;
-  height: 100%;
-  background-color: ${theme.colors.background.white};
-  box-shadow: -2px 0 20px rgba(0, 0, 0, 0.2);
-  z-index: 10;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  overflow: auto;
-
-  & h2 {
-    ${theme.typography.size_18}
-    font-weight: 600;
-  }
-
-  & h2::after {
-    content: '';
-    display: block;
-    width: 100%;
-    height: 1px;
-    background-color: ${theme.colors.gray.gray100};
-    margin-top: 8px;
-  }
-`;
-
-export const headerStyle = (theme: Theme) => css`
-  ${theme.typography.size_24}
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-top: 24px;
-  padding: 0 16px;
-`;
-
-const squadInfoStyle = css`
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  margin: 24px 0;
-  padding: 0 16px;
-
-  & h3 {
-    font-weight: 400;
-  }
-
-  & div {
-    display: flex;
-    gap: 4px;
-    align-items: center;
-  }
-`;
-
-const membersStyle = css`
-  padding: 0 16px;
-  margin-bottom: 24px;
-
-  & h2 {
-    margin-bottom: 12px;
-  }
-`;
-
-const settingsStyle = (theme: Theme) => css`
-  padding: 0 16px;
-  flex: 1;
-  margin-bottom: 24px;
-
-  & h2 {
-    margin-bottom: 12px;
-  }
-
-  & li {
-    padding: 8px 0;
-    cursor: pointer;
-  }
-
-  & li:hover {
-    opacity: 0.5;
-  }
-
-  & button {
-    text-align: left;
-    ${theme.typography.size_16}
-  }
-`;
-
-const exitButtonStyle = (theme: Theme) => css`
-  background-color: ${theme.colors.background.white};
-  ${theme.typography.size_14}
-  color: ${theme.colors.gray.gray200};
-  border-top: 1px solid ${theme.colors.gray.gray200};
-  font-weight: 700;
-  margin-top: auto;
-  height: 56px;
-  text-align: left;
-  padding: 12px 16px;
-`;
+export default Sidebar;
